Convert to Node.js
Modified the file structure to work with node.js
Includes the node.js server files
Server.js is the main server file, run with node.js to host the server.
Put the css and js files in their own folder, modified the html files to
aorrespond to these new locations.