Doctors now link to Listings
Doctors listed by index.html are now a link (unfortunately this messes
up the formatting, im happy to change it back to not being a <a></a>
class but for now it provides functionality)

Clicking the doctors link to the listing page and listing.js recives the
data from the databse but doesn't use it yet.

Listing.html isnt working well with the style and I'm not sure why. Even
the most recent git verison does it.